chore: remove conversion of created_utc to human readable datetime from get_response_from_pipeline
Previously, we had to convert the timestamp to a human readable datetime because our final response was dependent on the output of get_response_from_pipeline.

With the switch to MCP, a final response is generated before get_response_from_pipeline even executes.

Hence, the conversion to human readable datetime can be removed. Also, the get_response_from_pipeline function's only role should be to retrieve the relevant documents from MongoDB so that we can display these documents to the user.